% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_pack_row.r
\name{auto_pack_rows}
\alias{auto_pack_rows}
\title{Auto Pack Rows}
\usage{
auto_pack_rows(x, grouping_column, col.names)
}
\arguments{
\item{x}{data.frame. Data table.}

\item{grouping_column}{char. Name of the column to be used for grouping the rows.}

\item{col.names}{char. Vector containing the names of the columns to be displayed in the generated markdown report.}
}
\value{
A kableExtra object with rows grouped.
}
\description{
This function automates the grouping of rows with the same name to enhance
data visualization in tables using the \code{kbl()} function from the \code{kableExtra} package in rMarkdown reports.
}
\details{
This function was designed to enhance the visualization of data quality analysis in rMarkdown reports.
For more details of use, see the vignettes in the package documentation.
This function is designed to be used with the \code{kbl()} function from the \code{kableExtra} package.
}
\examples{
\dontrun{
data <- data.frame(Name = c('A', 'A', 'B', 'B', 'C'),
                   Value = c(1, 2, 3, 4, 5))
auto_pack_rows(data, 'Name', col.names = c('Value'))
}
}
