% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasus.R
\name{download_datasus}
\alias{download_datasus}
\title{Download and process DATASUS .dbc files}
\usage{
download_datasus(
  icd_list = NULL,
  fu_list = "all",
  selected_files,
  database,
  min_age = 0,
  max_age = Inf,
  proc_code = NA,
  export_file_code = "default",
  output_dir = NULL
)
}
\arguments{
\item{icd_list}{Character. List of ICD codes used for data filtering. Default is \code{NULL} to include all ICD codes.}

\item{fu_list}{Character. List of federative units (FUs) for extraction. Use 'all' to include all 27 federative units.
Default is 'all'.}

\item{selected_files}{Character. Vector containing the names of .dbc files to be extracted from the FTP server.}

\item{database}{Character. Acronym of the database for download. Acceptable values are 'AR', 'AQ', or 'RD'.}

\item{min_age}{Integer. Minimum age at the time of attendance for data filtering. Default is \code{0}.}

\item{max_age}{Integer. Maximum age at the time of attendance for data filtering. Default is \code{Inf}.}

\item{proc_code}{Character. Procedure code vector for data filtering. Default is \code{NA}.}

\item{export_file_code}{Character. String to be used for naming the output files. Use 'default' to include ICD name in
the file name. If there is more than one ICD, they will be separated by an underscore ('_').}

\item{output_dir}{Character. Output directory where .csv files will be saved. If \code{NULL}, the function will use the
current working directory. Default is \code{NULL}.}
}
\value{
Saves .csv files in the folder named 'datasus_files', which is created in the output directory.
}
\description{
This function extracts .dbc files from DATASUS, filters the data by provided ICD codes,
and saves the resulting table in .csv format in a local directory.
}
\details{
\itemize{
\item If \code{fu_list} is 'all', the function will download data from all 27 federative units of Brazil.
\item The files will be saved into the 'datasus_files' folder created in the specified output directory.
}
}
\examples{
\dontrun{
# Download and process data for ICD 'C00', federative unit 'SP', and save in current directory
download_datasus(
  icd_list = 'C00', 
  fu_list = 'SP', 
  selected_files = c('ARSP0108.dbc', 'ARSP0208.dbc'), 
  database = 'AR', 
  min_age = 0, 
  max_age = 100, 
  proc_code = NA, 
  export_file_code = 'default', 
  output_dir = NULL
)
}
}
