% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_files.R
\name{fetch_files}
\alias{fetch_files}
\title{Fetch list of downloaded .csv files extracted from DATASUS}
\usage{
fetch_files(
  output_dir = NULL,
  database,
  export_file_code = "default",
  icd_list,
  first_year,
  last_year
)
}
\arguments{
\item{output_dir}{Character. Directory where the function will read the files from. Default is the current working directory.}

\item{database}{Character. Acronym of the database for download. Acceptable values are 'AR', 'AQ', or 'RD'.}

\item{export_file_code}{Character. String used in the filename of the output files of the \code{download_datasus()} function. Default is 'default'.}

\item{icd_list}{Character. List of ICD codes used for data filtering.}

\item{first_year}{Integer. Oldest year of the database for extraction.}

\item{last_year}{Integer. Most recent year of the database for extraction.}
}
\value{
A character vector containing the list of .csv files that were extracted from the DATASUS database.
}
\description{
This function searches for .csv files downloaded by state, ICD code, and year, and returns a list
of all files that match this string pattern, following the filename pattern of the \code{download_datasus()} function.
}
\details{
\itemize{
\item The function searches for a folder named 'datasus_files' in the specified directory.
\item If the \code{output_dir} parameter is \code{NULL}, the function will search for the 'datasus_files' folder in the current working directory.
}
}
\examples{
\dontrun{
# Fetch list of downloaded files for database 'AR', ICD 'C00', and years 2010 to 2020
file_list <- fetch_files(
  database = 'AR', 
  export_file_code = 'C00', 
  icd_list = c('C00'), 
  first_year = 2010, 
  last_year = 2020
)
print(file_list)
}
}
