% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasus.r
\name{get_datasus}
\alias{get_datasus}
\title{Get DATASUS Files}
\usage{
get_datasus(
  icd_list = NULL,
  fu_list = "all",
  export_file_code = "default",
  database,
  first_year,
  last_year,
  min_age = 0,
  max_age = Inf,
  output_dir = NULL,
  proc_code = NA,
  rm_ufs_files = TRUE
)
}
\arguments{
\item{icd_list}{Character. List of ICD codes used for data filtering. Default is \code{NULL} to include all ICD codes.}

\item{fu_list}{Character. Vector containing the acronyms of the states for download. Use 'all' to download data from all 27 federative units of Brazil. Default is \code{"all"}.}

\item{export_file_code}{Character. If 'default', the function will use the default naming behavior.}

\item{database}{Character. Name of the database to be used. Acceptable values are 'AR', 'AQ', or 'RD'.}

\item{first_year}{Integer. The first year of the data interval to be downloaded.}

\item{last_year}{Integer. The last year of the data interval to be downloaded.}

\item{min_age}{Integer. Minimum age at the time of attendance for data filtering. Default is \code{0}.}

\item{max_age}{Integer. Maximum age at the time of attendance for data filtering. Default is \code{Inf}.}

\item{output_dir}{Character. Working directory where files will be saved. If \code{NULL}, the function will use the 'datasus_files' folder in the current working directory.}

\item{proc_code}{Character. Procedure code vector for data filtering. Default is \code{NA}.}

\item{rm_ufs_files}{Logical. If \code{TRUE}, the function will remove the .csv files for each state after merging them into one table. Default is \code{TRUE}.}
}
\value{
\itemize{
\item .csv files saved in the 'datasus_files' folder within the output directory.
\item A \code{data.frame} containing the merged table with data extracted from DATASUS.
}
}
\description{
This function automates the process of fetching file names from the DATASUS FTP server, downloading data by state,
converting from .dbc to .csv format, retrieving the list of .csv data files by state, and merging all datasets into one table.
}
\details{
\itemize{
\item If \code{fu_list} is 'all', the function will download data from all 27 federative units of Brazil.
\item Files are saved in the 'datasus_files' folder, created in the specified or default output directory.
\item This function is a wrapper for the following functions: \code{retrieve_filelist()}, \code{download_datasus()}, \code{fetch_files()}, and \code{merge_datasus()}.
}
}
\examples{
\dontrun{
# Download and merge data for ICD 'C00', database 'AR', and years 2010 to 2020
get_datasus(
  icd_list = 'C00', 
  fu_list = 'all', 
  export_file_code = 'default', 
  database = 'AR', 
  first_year = 2010, 
  last_year = 2020, 
  output_dir = NULL
)
}
}
