% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_list.r
\name{get_match_list}
\alias{get_match_list}
\title{Get Match List for Possible Duplicates}
\usage{
get_match_list(apac_table, silent = FALSE, print_time = TRUE)
}
\arguments{
\item{apac_table}{DataFrame. Table of patients that will be used for deduplication, containing patient information and calculated birth year.}

\item{silent}{Logical. If \code{TRUE}, the function will suppress warning messages when a patient code appears in more than one group of matches. Default is \code{FALSE}.}

\item{print_time}{Logical. If \code{TRUE}, the function will display the total processing time in the console. Default is \code{TRUE}.}
}
\value{
DataFrame. A matrix indicating the rows of patients with the same key that were identified as potential duplicates.
}
\description{
This function identifies potential duplicate patient records in an APAC table based on a margin of error of plus or minus 1 year in the calculated birth year.
}
\details{
\itemize{
\item The function uses the calculated birth year (\code{ano_nasc_calc}) to identify potential duplicate records with a margin of error of ±1 year, as the provided age might have been recorded before or after the patient's birthday.
\item Each identifier (row position) should ideally appear in only one row of the final match matrix. If the same identifier appears in multiple rows, manual verification and adjustment are required to prevent duplicate records during deduplication steps, such as when using the \code{merge_cns()} function.
\item A warning is displayed if a record appears in multiple groups, unless \code{silent = TRUE}.
}
}
\examples{
\dontrun{
# Identify potential duplicates
match_list <- get_match_list(apac_table = apac_data, silent = FALSE, print_time = TRUE)
head(match_list)
}
}
