% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_description.r
\docType{data}
\name{icd_description}
\alias{icd_description}
\title{ICD Description Data}
\format{
A data frame with 14233 rows and 2 variables:
\describe{
\item{cd_cid}{Character. ICD code.}
\item{ds_cid}{Character. Description of the ICD code.}
}
}
\source{
World Health Organization (WHO) - International Classification of Diseases (ICD) 10th revision.
Available at \url{http://www2.datasus.gov.br/cid10/V2008/cid10.htm}
}
\usage{
icd_description
}
\description{
This dataset contains the description of ICD-10 (International Classification of Diseases - 10th revision) codes.
}
\examples{
# Example code on how to use the dataset
df <- OncoDatasusR::icd_description
head(df)
}
\keyword{datasets}
