% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_datasus.R
\name{merge_datasus}
\alias{merge_datasus}
\title{Merge Downloaded .csv Files Extracted from DATASUS}
\usage{
merge_datasus(
  file_list,
  output_dir = NULL,
  database,
  export_file_code = "default",
  icd_list = NA,
  silence = FALSE,
  first_year,
  last_year,
  rm_ufs_files = TRUE
)
}
\arguments{
\item{file_list}{Character vector. List of file names to be read.}

\item{output_dir}{Character. Directory to read the files from. If \code{NULL}, the function will use the 'datasus_files' folder created in the current working directory.}

\item{database}{Character. Acronym of the database for download. Acceptable values are 'AR', 'AQ', or 'RD'.}

\item{export_file_code}{Character. String used in the name of the output files from the \code{download_datasus()} function.
Use 'default' to include the ICD name in the file name. If more than one ICD is provided, they will be separated by an underscore ('_'). If no ICD is provided, the code for the file name will be 'all'.}

\item{icd_list}{Character. List of ICD codes used for data filtering. Default is \code{NA} to include all ICD codes.}

\item{silence}{Logical. If \code{TRUE}, suppresses console messages about reading and writing files. Default is \code{FALSE}.}

\item{first_year}{Numeric. Oldest year of the data extraction. This value is used for naming the saved file.}

\item{last_year}{Numeric. Most recent year of the data extraction. This value is used for naming the saved file.}

\item{rm_ufs_files}{Logical. If \code{TRUE}, removes the .csv files for each state after merging them into one table. Default is \code{TRUE}.}
}
\value{
A merged .csv file containing data from all specified files.
}
\description{
This function reads all .csv files with data extracted from DATASUS by state and merges them into one comprehensive table.
}
\details{
\itemize{
\item The function searches for a folder named 'datasus_files' in the specified directory or in the current working directory if \code{output_dir} is \code{NULL}.
\item The merged file is saved in the same directory as the source files with a standardized naming convention.
\item Encodes patient CNS identifiers to base64 format for 'AQ' and 'AR' databases.
}
}
\examples{
\dontrun{
# Merge .csv files for database 'AR' and years 2010 to 2020
combined_data <- merge_datasus(
  file_list = c('AR_SP_C00_2010_2020.csv'), 
  database = 'AR',
  export_file_code = 'C00', 
  icd_list = c('C00'), 
  first_year = 2010,
  last_year = 2020, 
  silence = TRUE, 
  rm_ufs_files = TRUE
)
}
}
