% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sih.r
\name{merge_sih}
\alias{merge_sih}
\title{Merge SIH Records and Count Unique Patients}
\usage{
merge_sih(
  sih_table,
  first_year = NULL,
  last_year = NULL,
  age_group = "all",
  sex = NULL,
  write = FALSE,
  output_dir = NULL,
  filename = NA,
  silent = FALSE,
  usecep = FALSE,
  run_diagnostics = TRUE
)
}
\arguments{
\item{sih_table}{DataFrame. Input data frame containing SIH data.}

\item{first_year}{Integer. The earliest year to include in the data. If \code{NULL}, the earliest year of attendance in the database will be used.}

\item{last_year}{Integer. The latest year to include in the data. If \code{NULL}, the latest year of attendance in the database will be used.}

\item{age_group}{Character. The age group to consider for filtering. Options are 'all' (default), 'adult', 'young_adult', and 'pediatric'.}

\item{sex}{Character. The sex to consider for filtering. If \code{NULL}, all records are considered. Options are 'M' and 'F'.}

\item{write}{Logical. If \code{TRUE}, saves the output to a .csv file. Default is \code{FALSE}.}

\item{output_dir}{Character. Directory path where the file will be saved. If \code{NULL}, the function will use the 'datasus_files' folder in the current working directory. Default is \code{NULL}.}

\item{filename}{Character. Name of the output file. Default is \code{NA}, which saves the file as "SIH_merged.csv".}

\item{silent}{Logical. If \code{TRUE}, suppresses console messages about file saving. Default is \code{FALSE}.}

\item{usecep}{Logical. If \code{TRUE}, uses the first 5 digits of the CEP (postal code) as part of the patient identification key. Default is \code{FALSE}.}
}
\value{
DataFrame. A deduplicated data frame summarizing the count of unique patients.
}
\description{
This function filters and deduplicates patient records in a data frame containing SIH (Hospital Information System) data.
It can optionally save the output to a .csv file.
}
\details{
\itemize{
\item The function creates a unique patient identifier (\code{cod_paciente}) based on the provided birth date, sex, and optionally the first 5 digits of the CEP.
\item If multiple records for the same patient are found, the record with the earliest admission date is retained.
\item Deduplication behavior can be adjusted by toggling the \code{usecep} parameter.
\item The \code{age_group} parameter specifies the age group to consider for filtering. The minimum and maximum ages for each group are:
\itemize{
\item 'all': 0-99 years;
\item 'adult': 18-99 years;
\item 'young_adult': 15-39 years;
\item 'pediatric': 0-18 years;
}
\item If \code{write = TRUE}, the output file is saved in the specified or default directory.
}
}
\examples{
\dontrun{
# Merge SIH data and count unique patients, saving the result to a file
result <- merge_sih(
  sih_table = sih_data, 
  first_year = 2008,
  last_year = 2020,
  age_group = 'adult',
  sex  = NULL,
  write = TRUE, 
  output_dir = "path/to/save/", 
  filename = "sih_merged.csv",
  silent = FALSE, 
  usecep = TRUE
)
head(result)
}
}
