% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_apac.r
\name{prepare_apac}
\alias{prepare_apac}
\title{Finalize Preparation of APAC Table}
\usage{
prepare_apac(
  apac_table,
  first_year = NULL,
  last_year = NULL,
  age_type = "age_diagnosis",
  age_group = "all",
  sex = NULL,
  write = TRUE,
  filename,
  output_dir = NULL,
  run_diagnostics = TRUE,
  silent = FALSE
)
}
\arguments{
\item{apac_table}{DataFrame. The APAC table containing patient data.}

\item{first_year}{Integer. The earliest year to include in the data. If \code{NULL}, the earliest year of attendance in the database will be used.}

\item{last_year}{Integer. The latest year to include in the data. If \code{NULL}, the latest year of attendance in the database will be used.}

\item{age_type}{Character. The type of age to consider for filtering. Options are 'age_diagnosis' (default) and 'age_attendance'.}

\item{age_group}{Character. The age group to consider for filtering. Options are 'all' (default), 'adult', 'young_adult', and 'pediatric'.}

\item{sex}{Character. The sex to consider for filtering. If \code{NULL}, all records are considered. Options are 'M' and 'F'.}

\item{write}{Logical. If \code{TRUE}, saves the processed table as a \code{.csv} file in the 'exported_data' folder. Default is \code{TRUE}.}

\item{filename}{Character. The name of the file to be saved.}

\item{output_dir}{Character. Directory where the 'exported_data' folder will be created. If \code{NULL}, it will be created in the current working directory.}

\item{run_diagnostics}{Logical. If \code{TRUE}, generates and saves diagnostic information about the data processing. Default is \code{TRUE}.}

\item{silent}{Logical. If \code{TRUE}, suppresses console messages about processing steps. Default is \code{FALSE}.}
}
\value{
\itemize{
\item If \code{run_diagnostics = TRUE}: Returns a list with two elements:
\itemize{
\item \code{APAC_table}: The processed APAC table.
\item \code{diag_list}: A data frame with diagnostic information about the processing steps.
}
\item If \code{run_diagnostics = FALSE}: Returns only the processed APAC table.
}
}
\description{
This function performs the final preparation of the APAC table, including filtering records by the year of diagnosis,
patient age, and removing inconsistencies between the service date and the diagnosis date.
}
\details{
\itemize{
\item Filters records by year, patient age, and consistency between service and diagnosis dates.
\item The \code{age_type} parameter determines whether to use the age at diagnosis or the age at attendance and year of diagnosis or year of attendance for filtering.
\item The \code{age_group} parameter specifies the age group to consider for filtering. The minimum and maximum ages for each group are:
\itemize{
\item 'all': 0-99 years;
\item 'adult': 18-99 years;
\item 'young_adult': 15-39 years;
\item 'pediatric': 0-18 years;
}
\item Diagnostic information includes counts of records removed at each processing step.
\item Outputs can be saved as a \code{.csv} file if \code{write = TRUE}.
}
}
\examples{
\dontrun{
# Finalize preparation of APAC table with diagnostics and save the result
result <- prepare_apac(
  apac_table = apac_data, 
  first_year = 2008, 
  last_year = 2020,
  age_type = 'age_diagnosis',
  age_group = 'all',
  write = TRUE, 
  filename = 'prepared_apac.csv',
  output_dir = 'path/to/save/', 
  run_diagnostics = TRUE, 
  silent = FALSE
)
head(result$APAC_table)
head(result$diag_list)
}
}
