% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_linkage.r
\name{prepare_linkage}
\alias{prepare_linkage}
\title{Prepare the APAC Table for Identifying Possible Duplicates}
\usage{
prepare_linkage(APAC_table, run_diagnostics = TRUE)
}
\arguments{
\item{APAC_table}{DataFrame. The APAC data table to be processed.}

\item{run_diagnostics}{Logical. Indicates whether diagnostic information will be generated. Default is \code{TRUE}.}
}
\value{
\itemize{
\item If \code{run_diagnostics = TRUE}: Returns a list with two elements:
\itemize{
\item \code{APAC_table}: The APAC data table after creating the unique patient key.
\item \code{key_var_NA_table}: A data frame with the count of NA values for the variables used in the key.
}
\item If \code{run_diagnostics = FALSE}: Returns the APAC table with the unique patient key.
}
}
\description{
This function prepares the APAC table for evaluating possible duplicates by creating a unique patient key
using the first 5 digits of the ZIP code, primary CID, sex, date of diagnosis, and calculated year of birth.
The function also provides a diagnostic table with the count of NA values for the variables used in the key if required.
}
\details{
\itemize{
\item The function creates a potential key for identifying possible duplicates, composed of primary CID, sex,
the first 5 digits of the ZIP code, date of diagnosis, and the calculated year of birth.
\item Missing values in the variables used for the key are replaced with a unique sequence for each missing value
to prevent them from being considered equal.
}
}
\examples{
\dontrun{
# Prepare APAC table for linkage with diagnostics
linkage_data <- prepare_linkage(
  APAC_table = apac_data, 
  run_diagnostics = TRUE
)
head(linkage_data$APAC_table)
head(linkage_data$key_var_NA_table)
}
}
