% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_patient_table.r
\name{prepare_patient_table}
\alias{prepare_patient_table}
\title{Prepare a Descriptive Table of Patient Characteristics}
\usage{
prepare_patient_table(
  df,
  method = "APAC",
  age_group = "adult",
  age_type = "age_diagnosis",
  split = NULL
)
}
\arguments{
\item{df}{DataFrame. Table where each row represents a patient.}

\item{method}{Character. Identifies the database used to create the table. Acceptable values are \code{'APAC'} or \code{'AIH'}.}

\item{age_group}{Character. Specifies the age group division. Options are \code{'all'}, \code{'adult'}, \code{'young_adult'}, or \code{'pediatric'}. See \strong{Details}.}

\item{age_type}{Character. Type of age to be used in the table. Options are \code{'age_diagnosis'} or \code{'age_attendance'}. Default is \code{'age_diagnosis'}.}

\item{split}{Character (optional). Variable to split the table. Options are \code{NULL}, \code{'icd'}, \code{'icd_group'}, \code{'procedure'}, or \code{'database'} (for AQ and AR). Default is \code{NULL}.}
}
\value{
DataFrame. A data frame containing the descriptive table of patients.
}
\description{
Prepares a descriptive table of patient characteristics from an APAC or AIH data table.
}
\details{
The function returns counts and percentages of patients by age group, sex, region, and clinical staging,
as well as the mean and standard deviation of age for the entire population.

\strong{Age groups are defined as follows:}
\itemize{
\item If \code{age_group = 'all'}: 0-17, 18-29, 30-49, 50-69, 70-99 years.
\item If \code{age_group = 'adult'}: 18-29, 30-49, 50-69, 70-99 years.
\item If \code{age_group = 'young_adult'}: 15-19, 20-24, 25-29, 30-34, 35-39 years.
\item If \code{age_group = 'pediatric'}: 0-1, 2-5, 6-11, 12-18 years.
}

In \code{age_type}, \code{'age_diagnosis'} uses calculated age at diagnosis, while \code{'age_attendance'} uses the age at the time of attendance registered in APAC and AIH files. Age at diagnosis is available only in APAC files.

This function is a wrapper for internal functions \verb{.\\calculate_mean_sd()} and \verb{.\\frequency_table_summary()} used by the package.
}
\examples{
\dontrun{
# Prepare a descriptive patient table with age groups 'all'
patient_table <- prepare_patient_table(
  df = apac_data, 
  method = 'APAC', 
  age_group = 'all', 
  split = 'icd'
)
head(patient_table)
}
}
