% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_apac.r
\name{read_apac}
\alias{read_apac}
\title{Read and Process APAC Records}
\usage{
read_apac(
  read_bases = c("AQ", "AR"),
  aq_file = NA,
  ar_file = NA,
  dir = NULL,
  first_year = NA,
  last_year = NA,
  file_code = NA,
  procedures_filename = NA,
  run_diagnostics = TRUE,
  select_variables = NA,
  clear_cache = TRUE,
  silent = FALSE,
  export_procedure_table = FALSE,
  selected_med_procedures,
  med_proced_to_remove,
  group_icd = FALSE,
  icd_group_file_path = NULL
)
}
\arguments{
\item{read_bases}{Character. Indicates which APAC bases will be extracted and analyzed. Options: \code{'AQ'}, \code{'AR'}, or both.}

\item{aq_file}{Character (optional). A string with the name of the AQ base file to be read.}

\item{ar_file}{Character (optional). A string with the name of the AR base file to be read.}

\item{dir}{Character. Directory containing .csv files of AQ/AR bases. If \code{NULL}, defaults to \code{'datasus_files'} in the current directory.}

\item{first_year}{Numeric (optional). The oldest year for the database. See \strong{Details}.}

\item{last_year}{Numeric (optional). The most recent year for the database. See \strong{Details}.}

\item{file_code}{Character (optional). A string used in naming output files generated by \code{get_datasus()} or \code{merge_datasus()}. Defaults to \code{'APAC'}.}

\item{procedures_filename}{Character (optional). Customizes the name of the file with the procedure table.}

\item{run_diagnostics}{Logical. If \code{TRUE}, performs diagnostic analysis. Default is \code{TRUE}.}

\item{select_variables}{Character (optional). Additional variables from the original APAC tables to retain in the processed table.}

\item{clear_cache}{Logical. If \code{TRUE}, frees RAM space during processing with \code{gc()}. Default is \code{TRUE}.}

\item{silent}{Logical. If \code{TRUE}, suppresses console messages. Default is \code{FALSE}.}

\item{export_procedure_table}{Logical. If \code{TRUE}, saves the frequency table of procedures described in APAC. Default is \code{FALSE}.}

\item{selected_med_procedures}{Character vector of SIGPTAP procedure codes to include in the analysis.}

\item{med_proced_to_remove}{Character vector of SIGPTAP procedure codes to exclude from the analysis.}

\item{group_icd}{Logical. If \code{TRUE}, groups procedures by ICD.}

\item{icd_group_file_path}{Character (optional). Path and name of the ICD grouping file. Must include columns \code{'icd'} and \code{'icd_group'}.}
}
\value{
\itemize{
\item If \code{run_diagnostics = TRUE}: Returns a list with:
\itemize{
\item \code{APAC}: The processed APAC table.
\item \code{diag_list}: A list of diagnostic tables, including errors and summaries.
}
\item If \code{run_diagnostics = FALSE}: Returns the processed APAC table.
}
}
\description{
This function reads the APAC database, combining the AQ and AR bases if present, selects variables of interest,
creates variables for year of service and diagnosis, exports the frequency table of procedures by ICD, and performs
a consistency diagnostic of the variables for subsequent analyses.
}
\details{
The \code{first_year} and \code{last_year} parameters define the time range for reading the AQ/AR base file unless otherwise specified.
These parameters are also part of the output procedure table's name.

The function identifies potential inconsistencies, including missing values, date formatting errors, and mismatched age or diagnostic codes.
}
\examples{
\dontrun{
# Read and process APAC data for AQ and AR bases, grouping by ICD
apac_data <- read_apac(
  read_bases = c('AQ', 'AR'), 
  aq_file = 'aq_2010_2020_full.csv',
  ar_file = 'ar_2010_2020_full.csv', 
  dir = 'datasus_files/', 
  first_year = 2010, 
  last_year = 2020, 
  file_code = 'default',
  run_diagnostics = TRUE, 
  export_procedure_table = TRUE, 
  selected_med_procedures = c('0304040029','0304040045'), 
  med_proced_to_remove = c('0304040070','0304040088'), 
  group_icd = TRUE, 
  icd_group_file_path = 'source_files/icd_group_cancer.csv'
)
head(apac_data$APAC)
}
}
