% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_filelist.R
\name{retrieve_filelist}
\alias{retrieve_filelist}
\title{Retrieve a List of Filenames from a DATASUS FTP Server}
\usage{
retrieve_filelist(database, first_year = 2008, last_year = Inf)
}
\arguments{
\item{database}{Character. Acronym of the database for download. Acceptable values are 'AR', 'AQ', or 'RD'.}

\item{first_year}{Integer. The oldest year of the database for extraction. Default is 2008.}

\item{last_year}{Integer. The most recent year of the database for extraction. Default is Inf.}
}
\value{
A Character vector containing the names of available files for download, filtered according to the provided parameters.
}
\description{
This function retrieves the names of files available for download on a DATASUS FTP server,
based on the provided database acronym, specified year range, and the maximum wait time to access the FTP.
}
\examples{
\dontrun{
# Retrieve file names from the 'AR' database, for the years 2010 to 2020
filelist <- retrieve_filelist(database = 'AR', first_year = 2010, last_year = 2020)
print(filelist)
}
}
