% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtap_procedures.r
\docType{data}
\name{sigtap_procedures}
\alias{sigtap_procedures}
\title{SIGTAP Procedures Data}
\format{
A data frame with 5494 rows and 2 variables:
\describe{
\item{key}{Character. SIGTAP procedure code}
\item{description}{Character. Description of the procedure}
...
}
}
\source{
SIGTAP - Sistema de Gerenciamento da Tabela de Procedimentos, Medicamentos e OPM do SUS. Competence year/month: 2024-11.
Available at \url{https://sigtap.datasus.gov.br/tabela-unificada/app/sec/inicio.jsp}
}
\usage{
sigtap_procedures
}
\description{
This dataset contains the description of the procedures code in the SIGTAP database.
}
\examples{
# Example code on how to use the dataset
df <- OncoDatasusR::sigtap_procedures
head(df)
}
\keyword{datasets}
