% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_apac.r
\name{summarise_apac}
\alias{summarise_apac}
\title{Summarize APAC Records}
\usage{
summarise_apac(apac_table, run_diagnostics = TRUE)
}
\arguments{
\item{apac_table}{Data.frame Data frame containing the APAC table to be used to summarize the information.}

\item{run_diagnostics}{Logical If TRUE, performs table diagnostic analyses and identifies potential inconsistencies.}

\item{age_type}{Character Type of age to be used in the analysis. Can be 'age_diagnosis' or 'age_attendance'. Default is 'age_diagnosis'.}
}
\value{
\itemize{
\item If \code{run_diagnostics} is TRUE, returns a list containing the summarized APAC table and diagnostic analyses.
\item If \code{run_diagnostics} is FALSE, returns the summarized APAC table.
}
}
\description{
This function summarizes the APAC table data by patient code, including details about the diagnoses and treatments,
and performs diagnostic analyses to identify potential inconsistencies.
}
\details{
If the \code{run_diagnostics} argument is TRUE, the output will be a list where the first object is the summarized APAC table.
Additional elements of the list are as follows:
\itemize{
\item \code{diagnostic_year_counts}: Counts the number of records with a diagnosis date earlier than 1998.
\item \code{apac_per_patient}: Table with the frequency of APACs by patient code.
\item \code{dt_table_diag}: Data frame with the count of the number of records with the same diagnosis date and year across different databases.
\item \code{dt_diag_unique_patient}: Proportion of patients with unique diagnoses relative to the total (\%).
\item \code{dt_diag_unique_apac}: Proportion of records with unique diagnoses relative to the total (\%).
\item \code{n_before_cut_age}: Number of patients before filtering by age at diagnosis.
\item \code{n_after_cut_age}: Number of patients after filtering by age at diagnosis.
\item \code{dt_diag_error_tbl}: Table containing information about diagnosis inconsistencies.
}

In \code{age_type}, 'age_diagnosis' uses calculated age at diagnosis, while 'age_attendance' uses the age at the time of attendance
registered in APAC and AIH files. Age at diagnosis is available only in APAC files.
}
\examples{
\dontrun{
# Summarize APAC data with diagnostic analyses
summarized_data <- summarise_apac(apac_table = apac_data, run_diagnostics = TRUE, max_age = 99)
head(summarized_data)
}
}
